requirements.txt
aiohttp==3.2.1
alabaster==0.7.10
async-timeout==3.0.0
attrs==18.1.0
Babel==2.5.3
blessings==1.6.1
certifi==2018.4.16
chardet==3.0.4
cmd2==0.9.3
docutils==0.14
future==0.16.0
idna==2.6
idna-ssl==1.0.1
imagesize==1.0.0
Jinja2==2.10
MarkupSafe==1.0
multidict==4.3.1
packaging==17.1
paho-mqtt==1.3.1
peewee==3.3.2
prettytable==0.7.2
psycopg2-binary==2.7.5
Pygments==2.2.0
PyMySQL==0.8.1
pyparsing==2.2.0
pyperclip==1.6.0
pytz==2018.4
requests==2.18.4
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
urllib3==1.22
wcwidth==0.1.7
yarl==1.2.4
